Switch to plain unittest
nose has been unmaintained for a long while, and the included
unittest module is plenty featureful to support everything the test
suite includes. Sadly this commit includes an awful lot of churn due to
whitespace changes only. This also significantly drops the use of global
variables in test files, makes use of pytest as the test runner, and
updates the python environment list in tox.ini.